requirements.txt
flask
flask-cors
openpyxl
Pillow
python-magic
